+++ sbt-scalajs-bundler/src/main/scala/scalajsbundler/sbtplugin/ScalaJSBundlerPlugin.scala
@@ -147,6 +147,10 @@ object ScalaJSBundlerPlugin extends AutoPlugin {
       * }}}
       *
       * The task returns the produced bundles.
+      * 
+      * The task is cached, so webpack is only launched when some of the
+      * used files have changed. The list of files to be monitored is
+      * provided by webpackMonitoredFiles
       *
       * @group tasks
       */
@@ -201,6 +205,39 @@ object ScalaJSBundlerPlugin extends AutoPlugin {
       settingKey[Boolean]("Whether webpack should emit source maps at all")
 
     /**
+      * Additional directories, monitored for webpack launch.
+      *
+      * Changes to files in these directories that match
+      * `includeFilter` scoped to `webpackMonitoredFiles` enable
+      *  webpack launch in `webpack` task.
+      *
+      * Defaults to an empty `Seq`.
+      * 
+      * @group settings
+      * @see [[webpackMonitoredFiles]]
+      */
+    val webpackMonitoredDirectories: SettingKey[Seq[File]] =
+      settingKey[Seq[File]]("Directories, monitored for webpack launch")
+
+    /**
+      * List of files, monitored for webpack launch.
+      * 
+      * By default includes the following files:
+      *  - Generated `package.json`
+      *  - Generated webpack config
+      *  - Custom webpack config (if any)
+      *  - Files, provided by `webpackEntries` task.
+      *  - Files from `webpackMonitoredDirectories`, filtered by
+      *    `includeFilter`
+      *
+      * @group settings
+      * @see [[webpackMonitoredDirectories]]
+      * @see [[webpack]]
+      */
+    val webpackMonitoredFiles: TaskKey[Seq[File]] =
+      taskKey[Seq[File]]("Files that trigger webpack launch")
+
+    /**
       * whether to enable the “reload workflow” for `webpack in fastOptJS`.
       *
       * When enabled, dependencies are pre-bundled so that the output of `fastOptJS` can directly
@@ -230,6 +267,60 @@ object ScalaJSBundlerPlugin extends AutoPlugin {
     val useYarn: SettingKey[Boolean] =
       settingKey[Boolean]("Whether to use yarn for updates")
 
+    /**
+      * Port, on which webpack-dev-server will be launched.
+      *
+      * Defaults to 8080.
+      *
+      * @see [[startWebpackDevServer]]
+      */
+    val webpackDevServerPort = SettingKey[Int](
+      "webpackDevServerPort",
+      "Port, on which webpack-dev-server operates"
+    )
+
+    /**
+      * Additional arguments to webpack-dev-server.
+      * 
+      * Defaults to an empty list.
+      *
+      * @see [[startWebpackDevServer]]
+      */
+    val webpackDevServerExtraArgs = SettingKey[Seq[String]](
+      "webpackDevServerExtraArgs",
+      "Custom arguments to webpack-dev-server"
+    )
+
+    /**
+      * Start background webpack-dev-server process.
+      *
+      * If webpack-dev-server is already running, it will be restarted.
+      *
+      * The started webpack-dev-server receives the following arguments:
+      * - `--config` is set to value of `webpackConfigFile` setting.
+      * - `--port` is set to value of `webpackDevServerPort` setting.
+      * - Contents of `webpackDevServerExtraArgs` setting.
+      *
+      * @see [[stopWebpackDevServer]]
+      * @see [[webpackDevServerPort]]
+      * @see [[webpackDevServerExtraArgs]]
+      */
+    val startWebpackDevServer = TaskKey[Unit](
+      "startWebpackDevServer",
+      "(Re)start webpack-dev-server process"
+    )
+
+    /**
+      * Stop running webpack-dev-server process.
+      *
+      * Does nothing if the server is not running.
+      *
+      * @see [[startWebpackDevServer]]
+      */
+    val stopWebpackDevServer = TaskKey[Unit](
+      "stopWebpackDevServer",
+      "Stop webpack-dev-server process (if running)"
+    )
   }
 
   private val scalaJSBundlerPackageJson =
@@ -237,6 +328,12 @@ object ScalaJSBundlerPlugin extends AutoPlugin {
 
   private val scalaJSBundlerWebpackConfig =
     TaskKey[File]("scalaJSBundlerWebpackConfig", "Write the webpack configuration file", KeyRanks.Invisible)
+
+  private val webpackDevServer = SettingKey[WebpackDevServer](
+    "webpackDevServer",
+    "Global WebpackDevServer instance",
+    KeyRanks.Invisible
+  )
 
   private[scalajsbundler] val ensureModuleKindIsCommonJSModule =
     SettingKey[Boolean](
@@ -265,8 +362,28 @@ object ScalaJSBundlerPlugin extends AutoPlugin {
     ensureModuleKindIsCommonJSModule := {
       if (scalaJSModuleKind.value == ModuleKind.CommonJSModule) true
       else sys.error(s"scalaJSModuleKind must be set to ModuleKind.CommonJSModule in projects where ScalaJSBundler plugin is enabled")
-    }
+    },
 
+    // Make these settings project-level, since we don't expect much
+    // difference between configurations/stages. This way the
+    // API user can modify it just once.
+    webpackMonitoredDirectories := Seq(),
+    (includeFilter in webpackMonitoredFiles) := AllPassFilter,
+
+    // The defaults are specified at top level.
+    webpackDevServerPort := 8080,
+    webpackDevServerExtraArgs := Seq(),
+
+    // We can only have one server per project - for simplicity
+    webpackDevServer := new WebpackDevServer(),
+
+    (onLoad in Global) := {
+      (onLoad in Global).value.compose(
+        _.addExitHook {
+          webpackDevServer.value.stop()
+        }
+      )
+    }
   ) ++
     inConfig(Compile)(perConfigSettings) ++
     inConfig(Test)(perConfigSettings ++ testSettings)
@@ -464,8 +581,74 @@ object ScalaJSBundlerPlugin extends AutoPlugin {
           streams.value.log
         ),
 
-      webpackEmitSourceMaps in stageTask := (emitSourceMaps in stageTask).value
+      webpackEmitSourceMaps in stageTask := (emitSourceMaps in stageTask).value,
 
+      webpackMonitoredFiles in stageTask := {
+        val generatedWebpackConfigFile = (scalaJSBundlerWebpackConfig in stageTask).value
+        val customWebpackConfigFile = (webpackConfigFile in stageTask).value
+        val packageJsonFile = scalaJSBundlerPackageJson.value
+        val entries = (webpackEntries in stageTask).value
+
+        val filter = (includeFilter in webpackMonitoredFiles).value
+        val dirs = (webpackMonitoredDirectories in stageTask).value
+
+        val additionalFiles = dirs.flatMap(
+          dir => (dir ** filter).get
+        )
+
+        packageJsonFile +:
+          generatedWebpackConfigFile +:
+          customWebpackConfigFile ++:
+          entries.map(_._2) ++:
+          // Entries only contain launchers - we need to monitor
+          // Scala.js bundles themselves, too.
+          stageTask.value.data +:
+          additionalFiles
+      },
+
+      // webpack-dev-server wiring
+      startWebpackDevServer in stageTask := {
+        val serverDir = installWebpackDevServer.value
+
+        // We need to execute the full webpack task once, since it generates
+        // the required config file
+        (webpack in stageTask).value
+
+        val port = (webpackDevServerPort in stageTask).value
+        val extraArgs = (webpackDevServerExtraArgs in stageTask).value
+
+        // This duplicates file layout logic from `Webpack`
+        val targetDir = (npmUpdate in stageTask).value
+        val customConfigOption = (webpackConfigFile in stageTask).value
+        val generatedConfig = (scalaJSBundlerWebpackConfig in stageTask).value
+
+        val config = customConfigOption match {
+          case Some(customConfig) => targetDir / customConfig.name
+          case None => generatedConfig
+        }
+
+        // To match `webpack` task behavior
+        val workDir = targetDir
+
+        // Server instance is project-level
+        val server = webpackDevServer.value
+        val logger = (streams in stageTask).value.log
+
+        server.start(
+          serverDir,
+          workDir,
+          config,
+          port,
+          extraArgs,
+          logger
+        )
+      },
+
+      // Stops the global server instance, but is defined on stage
+      // level to match `startWebpackDevServer`
+      stopWebpackDevServer in stageTask := {
+        webpackDevServer.value.stop()
+      }
     )
   }
 
@@ -476,8 +659,8 @@ object ScalaJSBundlerPlugin extends AutoPlugin {
       val targetDir = npmUpdate.value
       val generatedWebpackConfigFile = (scalaJSBundlerWebpackConfig in stage).value
       val customWebpackConfigFile = (webpackConfigFile in stage).value
-      val packageJsonFile = scalaJSBundlerPackageJson.value
       val entries = (webpackEntries in stage).value
+      val monitoredFiles = (webpackMonitoredFiles in stage).value
 
       val cachedActionFunction =
         FileFunction.cached(
@@ -492,12 +675,8 @@ object ScalaJSBundlerPlugin extends AutoPlugin {
             log
           ).to[Set]
         }
-      cachedActionFunction(Set(
-        generatedWebpackConfigFile,
-        packageJsonFile
-      ) ++
-        (webpackConfigFile in stage).value.map(Set(_)).getOrElse(Set.empty) ++
-        entries.map(_._2).to[Set] + stage.value.data).to[Seq] // Note: the entries should be enough, excepted that they currently are launchers, which do not change even if the scalajs stage output changes
+
+      cachedActionFunction(monitoredFiles.to[Set]).to[Seq]
     }
 
   /**
@@ -513,6 +692,30 @@ object ScalaJSBundlerPlugin extends AutoPlugin {
         log.info(s"Installing jsdom in ${installDir.absolutePath}")
         IO.createDirectory(installDir)
         Npm.run("install", "jsdom")(installDir, log)
+      }
+      installDir
+    }
+
+  /**
+    * Locally installs webpack-dev-server.
+    *
+    * @return Installation directory
+    */
+  lazy val installWebpackDevServer: Def.Initialize[Task[File]] =
+    Def.task {
+      val installDir = target.value / "scalajs-bundler-webpack-dev-server"
+      val log = streams.value.log
+      val webpackVersion = (version in webpack).value
+
+      if (!installDir.exists()) {
+        log.info(s"Installing webpack-dev-server in ${installDir.absolutePath}")
+        IO.createDirectory(installDir)
+        Npm.run(
+          "install",
+          // Webpack version should match the setting
+          "webpack@" + webpackVersion,
+          "webpack-dev-server"
+        )(installDir, log)
       }
       installDir
     }