+++ sbt-scalajs-bundler/src/main/scala/scalajsbundler/util/Commands.scala
@@ -4,7 +4,7 @@ import sbt._
 
 object Commands {
 
-  def run(cmd: String, cwd: File, logger: Logger): Unit = {
+  def run(cmd: Seq[String], cwd: File, logger: Logger): Unit = {
     val process = Process(cmd, cwd)
     val code = process ! logger
     if (code != 0) {