+++ sbt-scalajs-bundler/src/sbt-test/sbt-scalajs-bundler/webpack-dev-server/src/main/scala/example/Main.scala
@@ -1,0 +1,9 @@
+package example
+
+import scala.scalajs.js.JSApp
+
+object Main extends JSApp {
+  def main(): Unit = {
+    println("Hello")
+  }
+}